Add optional max length attribute to @NotBlank
By default this will produce 2 different error messages, one for "not blank" and another for "max length".

An alternative of a {avaje.NotBlank.max.message} was tried which combined the 2 messages into a single error message but this felt less satisfactory than the 2 messages which are more clear as to the reason of the violation error.