Change andThen() to continue false when first validator is false
This is fixing a semantic bug (that has no impact because the validation chain stops anyway) to return false indicating that the validation shouldn't continue.

It has no impact because currently the generated code uses a single chain and that DOES actually stop processing (but then just returns the wrong value).

Tidy this up so that a followup refactor turning this into a lambda is then no behaviour change.