Add support for sharing instances of NotNull NotBlank adapters
When these are not customised and use the default group and default message, we can use shared "default adapters" instead of creating a new adapter for each instance.

This will be beneficial (reduce memory consumption) for the case where lots of NotNull & NotBlank etc used without any customisation of the groups or message.

This approach can be extended to also apply to other common validators like Email, UUID, AssertTrue, Positive etc.